requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.2.5
contourpy==1.3.2
cycler==0.12.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
Faker==37.4.0
fonttools==4.58.4
holidays==0.75
importlib_resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
matplotlib==3.10.3
narwhals==1.43.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
prompt_toolkit==3.0.51
prophet==1.1.7
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.2.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
stanio==0.5.1
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
